# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Edo Talent Hunt API",
    "version": "1.0.0",
    "description": "A comprehensive authentication API for Edo Talent Hunt with email verification, JWT authentication, and password management.",
    "contact": {
      "name": "Edo Talent Hunt Team",
      "email": "tolu.ajuwon@aceall.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://eth-project-backend-1086159474664.europe-west1.run.app",
      "description": "Production server (Google Cloud Run)"
    },
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme"
      }
    },
    "schemas": {
...
      "AdminRegistration": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "registrationNumber": {
            "type": "string"
          },
          "registrationType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "userId": {
            "type": "object",
            "properties": {
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "email": {
                "type": "string"
              }
            }
          },
          "personalInfo": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AdminTransaction": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "userId": {
            "type": "object"
          },
          "registrationId": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
  "paths": {
    "/api/v1/admin/registrations": {
      "get": {
        "summary": "Get all registrations (Admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "submitted",
                "approved",
                "rejected"
              ]
            },
            "description": "Filter by registration status"
          },
          {
            "in": "query",
            "name": "registrationType",
            "schema": {
              "type": "string",
              "enum": [
                "individual",
                "group"
              ]
            },
            "description": "Filter by registration type"
          },
          {
            "in": "query",
            "name": "registrationNumber",
            "schema": {
              "type": "string"
            },
            "description": "Filter by registration number (partial match)"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter registrations from this date"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter registrations until this date"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search in registration number, name, or email"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "default": "createdAt"
            },
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "All registrations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "registrations": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/AdminRegistration"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "currentPage": {
                              "type": "integer"
                            },
                            "totalPages": {
                              "type": "integer"
                            },
                            "totalCount": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "hasNextPage": {
                              "type": "boolean"
                            },
                            "hasPrevPage": {
                              "type": "boolean"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/admin/transactions": {
      "get": {
        "summary": "Get all payment transactions (Admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "initiated",
                "pending",
                "successful",
                "failed",
                "cancelled",
                "refunded"
              ]
            },
            "description": "Filter by transaction status"
          },
          {
            "in": "query",
            "name": "paymentMethod",
            "schema": {
              "type": "string"
            },
            "description": "Filter by payment method"
          },
          {
            "in": "query",
            "name": "currency",
            "schema": {
              "type": "string",
              "default": "NGN"
            },
            "description": "Filter by currency"
          },
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by user ID"
          },
          {
            "in": "query",
            "name": "registrationId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by registration ID"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter transactions from this date"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter transactions until this date"
          },
          {
            "in": "query",
            "name": "amountMin",
            "schema": {
              "type": "number"
            },
            "description": "Minimum amount filter"
          },
          {
            "in": "query",
            "name": "amountMax",
            "schema": {
              "type": "number"
            },
            "description": "Maximum amount filter"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search in transaction reference or gateway reference"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "default": "createdAt"
            },
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "All payment transactions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transactions": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/AdminTransaction"
                          }
                        },
                        "pagination": {
                          "type": "object"
                        },
                        "summary": {
                          "type": "object",
                          "properties": {
                            "totalAmount": {
                              "type": "number"
                            },
                            "successfulTransactions": {
                              "type": "integer"
                            },
                            "failedTransactions": {
                              "type": "integer"
                            },
                            "pendingTransactions": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "summary": "Get all users (Admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "contestant",
                "admin",
                "judge"
              ]
            },
            "description": "Filter by user role"
          },
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by active status"
          },
          {
            "in": "query",
            "name": "isEmailVerified",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by email verification status"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter users from this date"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter users until this date"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search in name or email"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "default": "createdAt"
            },
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "All users retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/admin/dashboard": {
      "get": {
        "summary": "Get dashboard statistics (Admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "registrations": {
                          "type": "object",
                          "properties": {
                            "byStatus": {
                              "type": "array"
                            },
                            "byType": {
                              "type": "array"
                            },
                            "total": {
                              "type": "integer"
                            }
                          }
                        },
                        "payments": {
                          "type": "object",
                          "properties": {
                            "byStatus": {
                              "type": "array"
                            },
                            "totalAmount": {
                              "type": "number"
                            },
                            "totalTransactions": {
                              "type": "integer"
                            }
                          }
                        },
                        "users": {
                          "type": "object",
                          "properties": {
                            "byRole": {
                              "type": "array"
                            },
                            "total": {
                              "type": "integer"
                            }
                          }
                        },
                        "recent": {
                          "type": "object",
                          "properties": {
                            "registrations": {
                              "type": "array"
                            },
                            "transactions": {
                              "type": "array"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/v1/registrations": {
      "get": {
        "summary": "Get user's registrations",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Registrations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Registration"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new registration",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "registrationType"
                ],
                "properties": {
                  "registrationType": {
                    "type": "string",
                    "enum": [
                      "individual",
                      "group"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration created successfully"
          },
          "400": {
            "description": "User already has an active registration"
          }
        }
      }
    },
...
  }
}